UI: move gameEdit to its own component
