+++ client/src/components/GameForm/GameForm.jsx
@@ -35,10 +35,13 @@ export default class GameForm extends React.Component {
     } = this.props.game
 
     const { editMode, errors } = this.props
+
+    // TODO : make title editable 
     return (
       <div>
         <h1>
-          <EditableText
+          {title}
+          {/* <EditableText
             type="input"
             text={title}
             error={errors["title"]}
@@ -49,7 +52,7 @@ export default class GameForm extends React.Component {
               game.title = d
               this.updateGame( game );
             }}
-            />
+            /> */}
         </h1>
 
           {/*
