+++ client/src/roles.js
@@ -1,0 +1,30 @@
+
+let contributor = {
+  "add_game" : true,
+  "edit_game" : true,
+  "delete_own_game" : false,
+  "delete_others_game" : false
+}
+
+let acl = {
+  contributor
+}
+
+export function isAuthorized(action, user) {
+
+  // unlogged user has no rights
+  console.log(user);
+  if (!user.is_auth)
+    return false
+
+  let roles = user.roles
+
+  // superuser have all rights
+  if (roles.indexOf("superuser") > -1) return true
+  else {
+    if (roles.indexOf("editor") > -1)
+      return acl["editor"][action]
+    else
+      return acl["contributor"][action]
+  }
+}
